Rename /shared/ui to /shared/components and move shared/ui/layout to /shared/layouts
