+++ application/account-management/WebApp/routes/register/expired.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { HorizontalHeroLayout } from "@/shared/ui/layout/HorizontalHeroLayout";
-import { ErrorMessage } from "@/shared/ui/components/ErrorMessage";
+import { HorizontalHeroLayout } from "@/shared/layouts/HorizontalHeroLayout";
+import { ErrorMessage } from "@/shared/components/ErrorMessage";
 import Timeout from "@spectrum-icons/illustrations/Timeout";
 import { Link } from "@repo/ui/components/Link";
 import { useRegistration } from "./-shared/actions";
