+++ application/account-management/WebApp/routes/register/verify.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { HorizontalHeroLayout } from "@/shared/ui/layout/HorizontalHeroLayout";
-import { ErrorMessage } from "@/shared/ui/components/ErrorMessage";
+import { HorizontalHeroLayout } from "@/shared/layouts/HorizontalHeroLayout";
+import { ErrorMessage } from "@/shared/components/ErrorMessage";
 import { Trans } from "@lingui/macro";
 import { Navigate } from "@tanstack/react-router";
 import { Button } from "@repo/ui/components/Button";
