+++ application/account-management/WebApp/shared/components/HeroImage.tsx
