+++ application/account-management/WebApp/shared/components/oneTimePassword/DigitPattern.tsx
